zebra: Address invalid memory access on VRF delete
A new 'zvrf' is created when the VRF is created.
This happens in zebra_vrf_new().
The rib_table_info_t points to this 'zvrf'.

During VRF deletion, the 'zvrf' is deleted.
However, the rib_table_info_t continues to hold a dangling
pointer to the deleted 'zvrf'.
Further, as part of route delete messages to the FPM, this dangling
pointer is accessed. ]
The table-id fetched from this dangling pointer will be corrupt.

Prevent 'zvrf' access completely in the FPM while sending route add
and delete messages. The 'zvrf' is only required to fetch the table-id.
Instead of accessing the table-id  via 'zvrf', store the table-id in
rib_table_info_t directly and access it.

Signed-off-by: NaveenThanikachalam <nthanikachal@vmware.com>